+++ src/condition/hold.rs
@@ -51,6 +51,12 @@ impl Hold {
         self.time_kind = kind;
         self
     }
+
+    /// Returns the associated timer.
+    #[must_use]
+    pub fn timer(&self) -> &Timer {
+        &self.timer
+    }
 }
 
 impl InputCondition for Hold {