+++ src/condition/tap.rs
@@ -43,6 +43,12 @@ impl Tap {
         self.time_kind = kind;
         self
     }
+
+    /// Returns the associated timer.
+    #[must_use]
+    pub fn timer(&self) -> &Timer {
+        &self.timer
+    }
 }
 
 impl InputCondition for Tap {