+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- New `Cooldown` input condition.
+- `Cooldown` input condition.
 
 ## [0.17.0] - 2025-08-18